requirements.txt
kafka-python-ng
pyspark==3.5.0
